refactor: move label file reconstruction logic to labels service for better organization
