+++ apps/backend/src/routes/labels.routes.ts
@@ -14,6 +14,7 @@ import {
   updateLabel,
   deleteLabel,
   getLabelCharacters,
+  reconstructFileForLabel,
   type PublicLabel,
   type LabelDetail,
   type ListLabelsFilters,
@@ -51,7 +52,6 @@ import {
   characters,
 } from "../db/schema/index.js";
 import { eq, asc, inArray, isNull, and, sql } from "drizzle-orm";
-import { reconstructRPYFile } from "../services/rpy-parser.service.js";
 import { calculateDialogueHash } from "../lib/hash.js";
 import { updateAuditFields } from "../lib/audit.js";
 import { calculateContentHash } from "../lib/hash.js";
@@ -94,127 +94,6 @@ interface UpdateLabelDialogueResponse {
 
 interface LabelCharactersResponse {
   characters: LabelCharacterWithInfo[];
-}
-
-/**
- * Generic type for database query operations shared by both db connections
- * and transactions. This allows the same function to work with either context.
- *
- * Only includes the query methods actually used by reconstructFileForLabel.
- */
-type QueryContext = {
-  select: ReturnType<typeof getDb>["select"];
-};
-
-/**
- * Reconstruct file content for a project file by fetching all labels
- * and their associated dialogue lines, then rebuilding the RPY file.
- *
- * References: db (getDb), labels, labelLines, characters, projectFiles tables,
- * and reconstructRPYFile service.
- */
-async function reconstructFileForLabel(
-  projectFileId: string,
-  db: QueryContext = getDb()
-): Promise<string> {
-  // Get the project file
-  const [projectFile] = await db
-    .select()
-    .from(projectFiles)
-    .where(eq(projectFiles.id, projectFileId))
-    .limit(1);
-
-  if (!projectFile) {
-    throw new NotFoundError("ProjectFile");
-  }
-
-  // `content` is the single source of truth for the current file state.
-  // `originalContent` is import-time baseline only and must not be used here.
-  const reconstructionBaseContent = projectFile.content;
-
-  // Fetch all labels for the project file
-  const allLabels = await db
-    .select({
-      id: labels.id,
-      labelName: labels.labelName,
-      title: labels.title,
-    })
-    .from(labels)
-    .where(
-      and(eq(labels.projectFileId, projectFile.id), isNull(labels.deletedAt))
-    )
-    .orderBy(asc(labels.labelPosition));
-
-  // If there are no labels, return current file content as-is
-  if (allLabels.length === 0) {
-    return reconstructRPYFile({
-      originalContent: reconstructionBaseContent,
-      updatedDialogue: new Map(),
-    });
-  }
-
-  // Build dialogue map for reconstruction
-  const updatedDialogue = new Map<
-    string,
-    Array<{ speaker: string | null; text: string }>
-  >();
-
-  // Batch fetch all label lines for all labels with speaker information
-  // Join with characters to get Ren'Py tag from speakerId
-  const allLabelLines = await db
-    .select({
-      labelId: labelLines.labelId,
-      speakerId: labelLines.speakerId,
-      speakerTag: characters.renpyTag,
-      content: labelLines.content,
-      sequence: labelLines.sequence,
-    })
-    .from(labelLines)
-    .leftJoin(characters, eq(labelLines.speakerId, characters.id))
-    .where(
-      and(
-        inArray(
-          labelLines.labelId,
-          allLabels.map((l) => l.id)
-        ),
-        isNull(labelLines.deletedAt)
-      )
-    )
-    .orderBy(asc(labelLines.sequence));
-
-  // Group lines by labelId in-memory
-  const linesByLabelId = new Map<
-    string,
-    Array<{ speaker: string | null; content: string }>
-  >();
-  for (const line of allLabelLines) {
-    if (!linesByLabelId.has(line.labelId)) {
-      linesByLabelId.set(line.labelId, []);
-    }
-    linesByLabelId.get(line.labelId)!.push({
-      // Use Ren'Py speaker tag for script-safe reconstruction
-      speaker: line.speakerTag ?? null,
-      content: line.content,
-    });
-  }
-
-  // Build dialogue map from grouped lines
-  for (const l of allLabels) {
-    const labelName = l.labelName || l.title;
-    const labelLinesData = linesByLabelId.get(l.id) || [];
-
-    const labelDialogue = labelLinesData.map((line) => ({
-      speaker: line.speaker,
-      text: line.content,
-    }));
-    updatedDialogue.set(labelName, labelDialogue);
-  }
-
-  // Reconstruct and return file content using current file content as base
-  return reconstructRPYFile({
-    originalContent: reconstructionBaseContent,
-    updatedDialogue,
-  });
 }
 
 // ============================================================================
